refactor: pass only stats information to the stats section component instead of the whole pokemon object
